fix: convert git uris with colons properly
This fixes a problem with git uris like `git@github.com:foo` -
where the uri comprises of a colon instead of a slash. Prior to
this fix, the uri would be given a leading slash, but retain the
colon, so the url would be `git@github.com/:foo`. This fixes that
issue so the url becomes `git@github.com/foo`.